romp/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.4.0
Pillow>=8.3.0
tqdm>=4.62.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
onnxruntime-gpu>=1.9.0
norfair>=2.0.0
# one-euro-filter>=1.0.0
pyrender>=0.1.45
trimesh>=3.9.0
pyglet>=1.5.0
moderngl>=5.6.0
moderngl-window>=2.4.0
PyOpenGL-accelerate>=3.1.0